@@ -37,6 +37,8 @@ dependencies:
   firebase_database: ^10.4.0
   firebase_auth: ^4.16.0
 
+  english_words: ^4.0.0
+
   cupertino_icons: ^1.0.6
 
   image_picker: ^1.0.7
@@ -55,13 +57,17 @@ dependencies:
 
   timeago: ^3.0.2
 
-  file: ^6.1.4
+  file:
   
   url_launcher: ^6.2.4
+
+  flutter_spinkit: ^5.2.0
 
   video_player_platform_interface: ^6.2.2
 
   plugin_platform_interface: ^2.1.8
+
+
 
   persistent_bottom_nav_bar: ^5.0.2
 
@@ -71,13 +77,26 @@ dependencies:
 
   chewie: ^1.7.5
 
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.7.0
 
   uuid: ^4.3.3
   firebase_messaging: ^14.6.8
   flutter_local_notifications: ^15.1.1
   http: ^1.2.0
   rxdart: ^0.27.7
+
+  
+  nlp: ^0.0.0
+
+  chips_choice: ^3.0.1
+  
+  get: ^4.6.6
+  mocktail_image_network: ^1.1.0
+  path: ^1.9.0
+  http: ^1.2.1
+  video_thumbnail: ^0.5.3
+
+
 
 
 
@@ -94,6 +113,7 @@ dev_dependencies:
   fake_cloud_firestore: ^2.4.8
   firebase_auth_mocks: ^0.13.0
   firebase_database_mocks: ^0.6.1
+
 
 
 
@@ -123,7 +143,11 @@ flutter:
   assets:
     - assets/blank_pfp.png
     - assets/base_image.png
-
+    - assets/default_profile_photo.png
+    - assets/profile_photo_1.png
+    - assets/profile_photo_2.png
+    - assets/profile_photo_3.png
+    
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
 
