@@ -1,16 +1,29 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+
+import 'package:info_hub_app/analytics/analytics_base.dart';
+
+import 'package:info_hub_app/message_feature/admin_message_view.dart';
+
 import 'package:info_hub_app/patient_experience/admin_experience_view.dart';
+import 'package:info_hub_app/theme/theme_manager.dart';
 import 'package:info_hub_app/topics/create_topic.dart';
 import 'package:info_hub_app/ask_question/question_view.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 
 class AdminHomepage extends StatefulWidget {
   final FirebaseFirestore firestore;
+  final FirebaseAuth auth;
   final FirebaseStorage storage;
+  final ThemeManager themeManager;
   const AdminHomepage(
-      {super.key, required this.firestore, required this.storage});
+      {super.key,
+      required this.firestore,
+      required this.auth,
+      required this.storage,
+      required this.themeManager});
   @override
   _AdminHomepageState createState() => _AdminHomepageState();
 }
@@ -59,6 +72,8 @@ class _AdminHomepageState extends State<AdminHomepage> {
                           return CreateTopicScreen(
                             storage: widget.storage,
                             firestore: widget.firestore,
+                            auth: widget.auth,
+                            themeManager: widget.themeManager,
                           );
                         },
                       ),
@@ -110,6 +125,7 @@ class _AdminHomepageState extends State<AdminHomepage> {
                   builder: (BuildContext context) {
                     return AdminExperienceView(
                       firestore: widget.firestore,
+                      auth: widget.auth,
                     );
                   },
                 ),
@@ -119,7 +135,51 @@ class _AdminHomepageState extends State<AdminHomepage> {
                 children: [
                   Icon(Icons.book),
                   Text(
-                    'View experiences',
+                    'View Experiences',
+                    style: TextStyle(color: Colors.black),
+                  ),
+                ],
+              ),
+            ),
+            ElevatedButton(
+              onPressed: () => Navigator.of(context).push(
+                CupertinoPageRoute(
+                  builder: (BuildContext context) {
+                    return AnalyticsBase(
+                      firestore: widget.firestore,
+                      storage: widget.storage,
+                    );
+                  },
+                ),
+              ),
+              child: const Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Icon(Icons.analytics),
+                  Text(
+                    'View Analytics',
+                    style: TextStyle(color: Colors.black),
+                  ),
+                ],
+              ),
+            ),
+            ElevatedButton(
+              onPressed: () => Navigator.of(context).push(
+                CupertinoPageRoute(
+                  builder: (BuildContext context) {
+                    return MessageView(
+                      firestore: widget.firestore,
+                      auth: widget.auth,
+                    );
+                  },
+                ),
+              ),
+              child: const Column(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Icon(Icons.message),
+                  Text(
+                    'Message feature',
                     style: TextStyle(color: Colors.black),
                   ),
                 ],
