@@ -9,56 +9,124 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:info_hub_app/discovery_view/discovery_view.dart';
 import 'package:info_hub_app/settings/settings_view.dart';
+import 'package:info_hub_app/theme/theme_manager.dart';
 import 'package:persistent_bottom_nav_bar/persistent_tab_view.dart';
 import 'package:info_hub_app/home_page/home_page.dart';
+import 'package:info_hub_app/admin/admin_dash.dart';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:info_hub_app/theme/theme_constants.dart';
 
-class Base extends StatelessWidget {
+class Base extends StatefulWidget {
   FirebaseAuth auth;
   FirebaseFirestore firestore;
   FirebaseStorage storage;
+  ThemeManager themeManager;
   Base(
       {super.key,
       required this.auth,
       required this.storage,
-      required this.firestore});
+      required this.firestore,
+      required this.themeManager});
+
+  @override
+  State<Base> createState() => _BaseState();
+}
+
+class _BaseState extends State<Base> {
+  String currentUserRoleType = '';
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    getCurrentUserRoleType();
+  }
+
+  Future<void> getCurrentUserRoleType() async {
+    User? user = widget.auth.currentUser;
+    if (user != null) {
+      DocumentSnapshot snapshot =
+          await widget.firestore.collection('Users').doc(user.uid).get();
+
+      setState(() {
+        currentUserRoleType = snapshot['roleType'];
+      });
+    }
+  }
+
+  List<Widget> getScreenBasedOnUser() {
+    if (currentUserRoleType == 'admin') {
+      return [
+        AdminHomepage(
+          auth: widget.auth,
+          storage: widget.storage,
+          firestore: widget.firestore,
+          themeManager: widget.themeManager,
+        ),
+        DiscoveryView(
+          auth: widget.auth,
+          storage: widget.storage,
+          firestore: widget.firestore,
+        ),
+        SettingsView(
+          auth: widget.auth,
+          firestore: widget.firestore,
+          storage: widget.storage,
+          themeManager: widget.themeManager,
+        ),
+      ];
+    } else {
+      return [
+        HomePage(
+          auth: widget.auth,
+          storage: widget.storage,
+          firestore: widget.firestore,
+        ),
+        DiscoveryView(
+          auth: widget.auth,
+          storage: widget.storage,
+          firestore: widget.firestore,
+        ),
+        SettingsView(
+          auth: widget.auth,
+          firestore: widget.firestore,
+          storage: widget.storage,
+          themeManager: widget.themeManager,
+        ),
+      ];
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     // Bottom Navigation Bar
     List<Widget> buildScreens() {
-      return [
-        HomePage(
-          auth: auth,
-          storage: storage,
-          firestore: firestore,
-        ),
-        DiscoveryView(
-          auth: auth,
-          storage: storage,
-          firestore: firestore,
-        ), // Should be replaced with the genuine page
-        const SettingsView(), // Should be replaced with the genuine page
-      ];
+      List<Widget> screens = getScreenBasedOnUser();
+      return screens;
     }
 
     // Styling for Bottom Navigation Bar
     List<PersistentBottomNavBarItem> navBarsItems() {
       return [
         PersistentBottomNavBarItem(
-          icon: const Icon(Icons.home),
-          activeColorPrimary: Colors.blue,
-          inactiveColorPrimary: Colors.grey,
+          icon: const Icon(Icons.home_outlined),
+          activeColorPrimary: COLOR_PRIMARY_LIGHT,
+          inactiveColorPrimary: Theme.of(context).brightness == Brightness.light
+              ? Colors.black
+              : Colors.white,
         ),
         PersistentBottomNavBarItem(
-          icon: const Icon(Icons.search),
-          activeColorPrimary: Colors.blue,
-          inactiveColorPrimary: Colors.grey,
+          icon: const Icon(Icons.search_outlined),
+          activeColorPrimary: COLOR_PRIMARY_LIGHT,
+          inactiveColorPrimary: Theme.of(context).brightness == Brightness.light
+              ? Colors.black
+              : Colors.white,
         ),
         PersistentBottomNavBarItem(
-          icon: const Icon(Icons.settings),
-          activeColorPrimary: Colors.blue,
-          inactiveColorPrimary: Colors.grey,
+          icon: const Icon(Icons.settings_outlined),
+          activeColorPrimary: COLOR_PRIMARY_LIGHT,
+          inactiveColorPrimary: Theme.of(context).brightness == Brightness.light
+              ? Colors.black
+              : Colors.white,
         ),
       ];
     }
@@ -73,7 +141,9 @@ class Base extends StatelessWidget {
       screens: buildScreens(),
       items: navBarsItems(),
       confineInSafeArea: true,
-      backgroundColor: Colors.white, // Default is Colors.white.
+      backgroundColor: Theme.of(context).brightness == Brightness.light
+          ? Colors.white
+          : Colors.black,
       handleAndroidBackButtonPress: true, // Default is true.
       resizeToAvoidBottomInset:
           true, // This needs to be true if you want to move up the screen when keyboard appears. Default is true.
@@ -82,7 +152,9 @@ class Base extends StatelessWidget {
           true, // Recommended to set 'resizeToAvoidBottomInset' as true while using this argument. Default is true.
       decoration: NavBarDecoration(
         borderRadius: BorderRadius.circular(10.0),
-        colorBehindNavBar: Colors.white,
+        colorBehindNavBar: Theme.of(context).brightness == Brightness.light
+            ? Colors.white
+            : Colors.black,
       ),
       popAllScreensOnTapOfSelectedTab: true,
       popActionScreens: PopActionScreensType.all,
@@ -98,7 +170,7 @@ class Base extends StatelessWidget {
         duration: Duration(milliseconds: 200),
       ),
       navBarStyle:
-          NavBarStyle.style5, // Choose the nav bar style with this property.
+          NavBarStyle.style6, // Choose the nav bar style with this property.
     );
   }
 }
