@@ -6,8 +6,11 @@ import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 import 'package:info_hub_app/admin/admin_dash.dart';
+import 'package:info_hub_app/discovery_view/discovery_view.dart';
 import 'package:info_hub_app/home_page/home_page.dart';
 import 'package:info_hub_app/push_notifications/push_notifications.dart';
+import 'package:info_hub_app/theme/theme_constants.dart';
+import 'package:info_hub_app/theme/theme_manager.dart';
 import 'notifications/notification.dart' as custom;
 import 'registration/start_page.dart';
 import 'package:provider/provider.dart';
@@ -15,8 +18,11 @@ import 'package:info_hub_app/services/database.dart';
 import 'package:info_hub_app/notifications/notifications.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:info_hub_app/topics/view_topic.dart';
 
 final navigatorKey = GlobalKey<NavigatorState>();
+
+ThemeManager themeManager = ThemeManager();
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -65,7 +71,7 @@ Future<void> main() async {
   ));
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
   final FirebaseFirestore firestore;
   final FirebaseAuth auth;
   final FirebaseStorage storage;
@@ -76,12 +82,38 @@ class MyApp extends StatelessWidget {
       required this.storage});
 
   @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  // Theme data listener
+  themeListener() {
+    if (mounted) {
+      setState(() {});
+    }
+  }
+
+  @override
+  void initState() {
+    themeManager.addListener((themeListener));
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    themeManager.removeListener(themeListener);
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: [
         StreamProvider<List<custom.Notification>>(
           create: (_) => DatabaseService(
-                  auth: auth, firestore: firestore, uid: auth.currentUser!.uid)
+                  auth: widget.auth,
+                  firestore: widget.firestore,
+                  uid: widget.auth.currentUser!.uid)
               .notifications,
           initialData: const [],
         ),
@@ -89,16 +121,21 @@ class MyApp extends StatelessWidget {
       child: MaterialApp(
         navigatorKey: navigatorKey,
         home: StartPage(
-          firestore: firestore,
-          storage: storage,
-          auth: auth,
+          firestore: widget.firestore,
+          storage: widget.storage,
+          auth: widget.auth,
+          themeManager: themeManager,
         ),
         routes: {
           '/notifications': (context) => Notifications(
-                auth: auth,
-                firestore: firestore,
+                auth: widget.auth,
+                firestore: widget.firestore,
               ),
         },
+        theme: lightTheme,
+        darkTheme: darkTheme,
+        themeMode: themeManager.themeMode,
+
         // home: HomePage(auth: auth, firestore: firestore, storage: storage)
         // home: AdminHomepage(firestore: firestore, storage: storage),
       ),
