@@ -28,23 +28,34 @@ class _NotificationsState extends State<Notifications> {
 
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Colors.red[400],
         title: const Text('Notifications'),
       ),
       body: ListView.builder(
         itemCount: userNotifications.length,
         itemBuilder: (context, index) {
           final notification = userNotifications[index];
-          return Dismissible(
-            key: Key(notification.id),
-            child: NotificationCard(notification: notification),
-            onDismissed: (direction) {
-              DatabaseService(
-                      auth: widget.auth,
-                      firestore: widget.firestore,
-                      uid: widget.auth.currentUser!.uid)
-                  .deleteNotification(notification.id);
-            },
+          return Column(
+            children: [
+              Dismissible(
+                key: Key(notification.id),
+                child: NotificationCard(notification: notification),
+                onDismissed: (direction) {
+                  DatabaseService(
+                          uid: widget.auth.currentUser!.uid,
+                          auth: widget.auth,
+                          firestore: widget.firestore)
+                      .deleteNotification(notification.id);
+                },
+              ),
+              if (index != userNotifications.length - 1)
+                Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 16),
+                  child: Container(
+                    height: 1,
+                    color: Colors.grey,
+                  ),
+                ),
+            ],
           );
         },
       ),
