@@ -2,6 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:info_hub_app/theme/theme_manager.dart';
 import 'registration_screen.dart';
 import 'package:info_hub_app/login/login_screen.dart';
 
@@ -9,11 +10,13 @@ class StartPage extends StatelessWidget {
   final FirebaseFirestore firestore;
   final FirebaseAuth auth;
   final FirebaseStorage storage;
+  final ThemeManager themeManager;
   const StartPage(
       {super.key,
       required this.firestore,
       required this.auth,
-      required this.storage});
+      required this.storage,
+      required this.themeManager});
 
   @override
   Widget build(BuildContext context) {
@@ -84,6 +87,7 @@ class StartPage extends StatelessWidget {
                           firestore: firestore,
                           auth: auth,
                           storage: storage,
+                          themeManager: themeManager,
                         ),
                       ),
                     );
