@@ -7,14 +7,20 @@ class UserModel {
   final String roleType;
   final List<String> likedTopics;
   final List<String> dislikedTopics;
+  List<String>? draftedTopics;
+  bool hasOptedOutOfExperienceExpectations;
+  String?
+      selectedProfilePhoto; // This is a nullable variable, as the user may not have a profile photo
 
-  UserModel({
-    required this.uid,
-    required this.firstName,
-    required this.lastName,
-    required this.email,
-    required this.roleType,
-    required this.likedTopics,
-    required this.dislikedTopics,
-  });
+  UserModel(
+      {required this.uid,
+      required this.firstName,
+      required this.lastName,
+      required this.email,
+      required this.roleType,
+      required this.likedTopics,
+      required this.dislikedTopics,
+      this.hasOptedOutOfExperienceExpectations = false,
+      this.draftedTopics,
+      this.selectedProfilePhoto});
 }
