@@ -1,112 +1,202 @@
+import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
+import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:info_hub_app/profile_view/profile_view.dart';
 import 'package:info_hub_app/registration/user_model.dart';
 import 'package:info_hub_app/notifications/manage_notifications.dart';
 import 'package:info_hub_app/services/database.dart';
+import 'package:info_hub_app/settings/general_settings.dart';
+import 'package:info_hub_app/theme/theme_manager.dart';
+import 'package:info_hub_app/settings/saved/saved_page.dart';
+import 'package:info_hub_app/settings/drafts/drafts_page.dart';
 import 'package:provider/provider.dart';
-import 'package:info_hub_app/screens/privacy_base.dart';
+import 'package:info_hub_app/settings/privacy_base.dart';
 import 'package:persistent_bottom_nav_bar/persistent_tab_view.dart';
+import 'package:info_hub_app/registration/start_page.dart';
+import 'package:info_hub_app/settings/help_page.dart'; // Import the help page widget
 
 class SettingsView extends StatefulWidget {
-  const SettingsView({super.key});
+  final FirebaseAuth auth;
+  final FirebaseFirestore firestore;
+  final FirebaseStorage storage;
+  final ThemeManager themeManager;
+  const SettingsView(
+      {super.key,
+      required this.auth,
+      required this.firestore,
+      required this.storage,
+      required this.themeManager});
 
   @override
   State<SettingsView> createState() => _SettingsViewState();
 }
 
 class _SettingsViewState extends State<SettingsView> {
+  bool isAdmin = false;
+
+  @override
+  void initState() {
+    super.initState();
+    initializeAdminStatus();
+  }
+
+  Future<void> initializeAdminStatus() async {
+    await isAdminUser();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return MultiProvider(
+      providers: [
+        StreamProvider<List<UserModel>>(
+          create: (_) => DatabaseService(
+            uid: FirebaseAuth.instance.currentUser!.uid,
+            firestore: FirebaseFirestore.instance,
+            auth: FirebaseAuth.instance,
+          ).users,
+          initialData: [], // Initial data while waiting for Firebase data
+        ),
+      ],
+      child: Scaffold(
         appBar: AppBar(
           title: const Text("Settings"),
         ),
         body: ListView(
           children: [
-            Container(
-              // decoration: BoxDecoration(
-              //   border: Border.all(color: Colors.black)
-              // ),
-              child: const ListTile(
-                leading: CircleAvatar(
-                  radius: 30,
-                  foregroundColor: Color.fromRGBO(226, 4, 4, 0.612),
-                  backgroundImage: AssetImage('assets/blank_pfp.png'),
-                ),
-                title: Text("Username"),
-                subtitle: Text("Role"),
-              ),
-            ),
-            GestureDetector(
+            ListTile(
+              title: const Text("Account"),
               onTap: () {
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                      builder: (context) => ProfileView(
+                            firestore: widget.firestore,
+                            auth: widget.auth,
+                          )),
+                );
+              },
+            ),
+            ListTile(
+              title: const Text("General"),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                    builder: (context) => GeneralSettings(
+                      themeManager: widget.themeManager,
+                    ),
+                  ),
+                );
+              },
+            ),
+            ListTile(
+              title: const Text("Notifications"),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                    builder: (context) => ManageNotifications(
+                        firestore: widget.firestore, auth: widget.auth),
+                  ),
+                );
+              },
+            ),
+            const ListTile(
+              title: Text('History'),
+            ),
+            ListTile(
+              title: const Text("Privacy"),
+              onTap: () {
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                    builder: (context) => const PrivacyPage(),
+                  ),
+                );
+              },
+            ),
+            ListTile(
+              title: const Text('Saved Topics'),
+              onTap: () {
+                // Navigate to the saved topics page when tapped
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                      builder: (context) => SavedPage(
+                            auth: widget.auth,
+                            firestore: widget.firestore,
+                            storage: widget.storage,
+                          )),
+                );
+              },
+            ),
+            if (isAdmin)
+              ListTile(
+                key: const Key("drafts_tile"),
+                title: const Text('Topic Drafts'),
+                onTap: () {
+                  // Navigate to the saved topics page when tapped
+                  Navigator.push(
+                    context,
+                    CupertinoPageRoute(
+                        builder: (context) => DraftsPage(
+                              auth: widget.auth,
+                              firestore: widget.firestore,
+                              storage: widget.storage,
+                            )),
+                  );
+                },
+              ),
+            ListTile(
+              title: const Text('Help'),
+              onTap: () {
+                // Navigate to the HelpPage when tapped
+                Navigator.push(
+                  context,
+                  CupertinoPageRoute(builder: (context) => const HelpPage()),
+                );
+              },
+            ),
+            const AboutListTile(
+              applicationLegalese: 'Legalese',
+              applicationName: 'TEAM TODO',
+              applicationVersion: '1.0.0',
+              aboutBoxChildren: [
+                Text('Liver information hub for young people'),
+              ],
+            ),
+            ListTile(
+              title: const Text('Log Out'),
+              onTap: () {
+                widget.auth.signOut();
                 PersistentNavBarNavigator.pushNewScreen(
                   context,
-                  screen: ManageNotifications(
-                    firestore: FirebaseFirestore.instance,
-                    auth: FirebaseAuth.instance,
+                  screen: StartPage(
+                    firestore: widget.firestore,
+                    auth: widget.auth,
+                    storage: widget.storage,
+                    themeManager: widget.themeManager,
                   ),
                   withNavBar: false,
                 );
               },
-              child: Container(
-                // decoration: BoxDecoration(
-                //   border: Border.all(color: Colors.black)
-                // ),
-                child: const ListTile(
-                  leading: Icon(Icons.notifications),
-                  title: Text('Manage Notifications'),
-                ),
-              ),
-            ),
-            Container(
-                // decoration: BoxDecoration(
-                //   border: Border.all(color: Colors.black)
-                // ),
-                child: ListTile(
-              leading: Icon(Icons.privacy_tip),
-              title: Text('Manage Privacy Settings'),
-              onTap: () {
-                PersistentNavBarNavigator.pushNewScreen(
-                  context,
-                  screen: PrivacyPage(),
-                  withNavBar: false,
-                );
-              },
-            )),
-            Container(
-              // decoration: BoxDecoration(
-              //   border: Border.all(color: Colors.black)
-              // ),
-              child: const ListTile(
-                leading: Icon(Icons.history_outlined),
-                title: Text('History'),
-              ),
-            ),
-            Container(
-              // decoration: BoxDecoration(
-              //   border: Border.all(color: Colors.black)
-              // ),
-              child: const ListTile(
-                leading: Icon(Icons.help),
-                title: Text('Help'),
-              ),
-            ),
-            Container(
-              // decoration: BoxDecoration(
-              //   border: Border.all(color: Colors.black)
-              // ),
-              child: const AboutListTile(
-                icon: Icon(Icons.info),
-                applicationLegalese: 'Legalese',
-                applicationName: 'TEAM TODO',
-                applicationVersion: '1.0.0',
-                aboutBoxChildren: [
-                  Text('Liver information hub for young people')
-                ],
-              ),
             ),
           ],
-        ));
+        ),
+      ),
+    );
+  }
+
+  Future<void> isAdminUser() async {
+    User? user = widget.auth.currentUser;
+    if (user != null) {
+      DocumentSnapshot snapshot =
+          await widget.firestore.collection('Users').doc(user.uid).get();
+      setState(() {
+        isAdmin = snapshot['roleType'] == 'admin';
+      });
+    }
   }
 }
