@@ -1,21 +1,36 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
+import 'package:firebase_auth_mocks/firebase_auth_mocks.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:info_hub_app/admin/admin_dash.dart';
+
+import 'package:info_hub_app/analytics/analytics_base.dart';
+import 'package:info_hub_app/main.dart';
+
+import 'package:info_hub_app/message_feature/admin_message_view.dart';
+
+import 'package:info_hub_app/patient_experience/admin_experience_view.dart';
 import 'package:info_hub_app/topics/create_topic.dart';
 import 'package:info_hub_app/ask_question/question_view.dart';
 import 'package:firebase_storage_mocks/firebase_storage_mocks.dart';
 
 void main() {
   late FirebaseFirestore firestore = FakeFirebaseFirestore();
+  late MockFirebaseAuth auth = MockFirebaseAuth();
   late MockFirebaseStorage mockStorage = MockFirebaseStorage();
   late Widget adminWidget;
 
   setUp(() {
     firestore = FakeFirebaseFirestore();
+
     adminWidget = MaterialApp(
-      home: AdminHomepage(firestore: firestore, storage: mockStorage),
+      home: AdminHomepage(
+        firestore: firestore,
+        auth: auth,
+        storage: mockStorage,
+        themeManager: themeManager,
+      ),
     );
   });
 
@@ -43,6 +58,37 @@ void main() {
     await tester.tap(find.text('View Thread'));
     await tester.pumpAndSettle();
     //expect(find.byType(CreateTopicScreen), findsOneWidget);
+  });
+
+  testWidgets('Test view experiences button', (WidgetTester tester) async {
+    await tester.pumpWidget(adminWidget);
+    await tester.tap(find.text('View Experiences'));
+    await tester.pumpAndSettle();
+    expect(find.byType(AdminExperienceView), findsOneWidget);
+  });
+
+  testWidgets('Test view analytics button', (WidgetTester tester) async {
+    await tester.pumpWidget(adminWidget);
+    await tester.tap(find.text('View Analytics'));
+    await tester.pumpAndSettle();
+    expect(find.byType(AnalyticsBase), findsOneWidget);
+  });
+
+  testWidgets('Test view message feature button', (WidgetTester tester) async {
+    await auth.createUserWithEmailAndPassword(
+        email: 'admin@gmail.com', password: 'Admin123!');
+    String uid = auth.currentUser!.uid;
+    await firestore.collection('Users').doc(uid).set({
+      'email': 'admin@gmail.com',
+      'firstName': 'John',
+      'lastName': 'Doe',
+      'roleType': 'admin'
+    });
+
+    await tester.pumpWidget(adminWidget);
+    await tester.tap(find.text('Message feature'));
+    await tester.pumpAndSettle();
+    expect(find.byType(MessageView), findsOneWidget);
   });
 
   testWidgets('Add admin test', (WidgetTester tester) async {
