@@ -12,36 +12,63 @@ void main() {
   late MockFirebaseStorage storage;
   late CollectionReference topicsCollectionRef;
   late Widget discoveryViewWidget;
-
-  setUp(() {
+  setUp(() async {
     firestore = FakeFirebaseFirestore();
     auth = MockFirebaseAuth();
     storage = MockFirebaseStorage();
     topicsCollectionRef = firestore.collection('topics');
-
+    await auth.createUserWithEmailAndPassword(
+        email: 'test@tested.org', password: 'Password123!');
+    String uid = auth.currentUser!.uid;
+    await firestore.collection('Users').doc(uid).set({
+      'email': 'test@tested.org',
+      'firstName': 'James',
+      'lastName': 'Doe',
+      'roleType': 'Patient'
+    });
     topicsCollectionRef.add({
       'title': 'B test',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
     topicsCollectionRef.add({
       'title': 'D test',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
     topicsCollectionRef.add({
       'title': 'A test',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
     topicsCollectionRef.add({
       'title': 'C test',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
 
     discoveryViewWidget = MaterialApp(
@@ -54,30 +81,13 @@ void main() {
     await tester.pumpWidget(discoveryViewWidget);
 
     expect(find.byType(TextField), findsOneWidget);
-    expect(find.widgetWithIcon(IconButton, Icons.search), findsOneWidget);
   });
-
-  /*
-  testWidgets('DiscoveryView back button pops user out of search screen', (WidgetTester tester) async {
-    await tester.pumpWidget(discoveryViewWidget);
-
-    expect(find.byWidget(discoveryViewWidget), findsOneWidget);
-
-    final backButton = find.widgetWithIcon(IconButton, Icons.arrow_back);
-
-    await tester.tap(backButton);
-    await tester.pumpAndSettle();
-
-    expect(find.byWidget(discoveryViewWidget), findsNothing);
-
-  });
-  */
 
   testWidgets('DiscoveryView search button does nothing (is null)',
       (WidgetTester tester) async {
     await tester.pumpWidget(discoveryViewWidget);
 
-    final searchButton = find.widgetWithIcon(IconButton, Icons.search);
+    final searchButton = find.byType(TextField);
 
     await tester.tap(searchButton);
     await tester.pump();
@@ -92,18 +102,33 @@ void main() {
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
     topicsCollectionRef.add({
       'title': 'Multiple will show 2',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
     topicsCollectionRef.add({
       'title': 'Multiple will show 3',
       'description': 'this is a test',
       'articleLink': '',
       'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
     });
 
     await tester.pumpWidget(discoveryViewWidget);
@@ -165,25 +190,240 @@ void main() {
     expect((textFinders.at(3).evaluate().single.widget as Text).data, 'D test');
   });
 
-  testWidgets('Show Post Dialog Test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
+  testWidgets('DiscoveryView will display categories as toggle buttons',
+      (WidgetTester tester) async {
+    await firestore.collection('categories').add({'name': 'Gym'});
+    await firestore.collection('categories').add({'name': 'School'});
+    await firestore.collection('categories').add({'name': 'Smoking'});
+
     await tester.pumpWidget(discoveryViewWidget);
     await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Gym'));
+    await tester.pumpAndSettle();
+    expect(find.text('Gym'), findsOneWidget);
+
+    await tester.ensureVisible(find.text('School'));
+    await tester.pumpAndSettle();
+    expect(find.text('School'), findsOneWidget);
+
+    await tester.ensureVisible(find.text('Smoking'));
+    await tester.pumpAndSettle();
+    expect(find.text('Smoking'), findsOneWidget);
+  });
+
+  testWidgets('DiscoveryView will display topics specific to one category',
+      (WidgetTester tester) async {
+    await firestore.collection('categories').add({'name': 'Gym'});
+    await firestore.collection('categories').add({'name': 'School'});
+    await firestore.collection('categories').add({'name': 'Smoking'});
+
+    topicsCollectionRef.add({
+      'title': 'Gym topic should only show',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym'],
+      'tags': ['Patient']
+    });
+
+    topicsCollectionRef.add({
+      'title': 'Gym topic should only show 2',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym'],
+      'tags': ['Patient']
+    });
+
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    await tester.tap(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    Finder cardFinder = find.byType(Card);
+    expect(cardFinder, findsNWidgets(2));
+
+    expect(find.text('Gym topic should only show'), findsOneWidget);
+    expect(find.text('Gym topic should only show 2'), findsOneWidget);
+  });
+
+  testWidgets('Tapping the filters twice will turn it off',
+      (WidgetTester tester) async {
+    await firestore.collection('categories').add({'name': 'Gym'});
+    await firestore.collection('categories').add({'name': 'School'});
+    await firestore.collection('categories').add({'name': 'Smoking'});
+
+    topicsCollectionRef.add({
+      'title': 'Gym topic should only show',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym'],
+      'tags': ['Patient']
+    });
+
+    topicsCollectionRef.add({
+      'title': 'Gym topic should only show 2',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym'],
+      'tags': ['Patient']
+    });
+
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    await tester.tap(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    //filters to two
+    Finder cardFinder = find.byType(Card);
+    expect(cardFinder, findsNWidgets(2));
+
+    //pressing filter again
+    await tester.tap(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    //all topics are now visible as filter is off
+    expect(cardFinder, findsNWidgets(6));
+  });
+
+  testWidgets('DiscoveryView will display topics based on multiple filters',
+      (WidgetTester tester) async {
+    await firestore.collection('categories').add({'name': 'Gym'});
+    await firestore.collection('categories').add({'name': 'School'});
+    await firestore.collection('categories').add({'name': 'Smoking'});
+
+    topicsCollectionRef.add({
+      'title': 'Gym topic should only show',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym'],
+      'tags': ['Patient']
+    });
+
+    topicsCollectionRef.add({
+      'title': 'Gym and smoking',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Gym', 'Smoking'],
+      'tags': ['Patient']
+    });
+
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Gym'));
+    await tester.pumpAndSettle();
+    await tester.tap(find.text('Gym'));
+    await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Smoking'));
+    await tester.pumpAndSettle();
+    await tester.tap(find.text('Smoking'));
+    await tester.pumpAndSettle();
+
+    Finder cardFinder = find.byType(Card);
+    expect(cardFinder, findsNWidgets(1));
+
+    expect(find.text('Gym and smoking'), findsOneWidget);
+  });
+
+  testWidgets('DiscoveryView will display topics based on filter and search',
+      (WidgetTester tester) async {
+    await firestore.collection('categories').add({'name': 'Gym'});
+    await firestore.collection('categories').add({'name': 'School'});
+    await firestore.collection('categories').add({'name': 'Smoking'});
+
+    topicsCollectionRef.add({
+      'title': 'Smoking topic',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Smoking'],
+      'tags': ['Patient']
+    });
+
+    topicsCollectionRef.add({
+      'title': 'Smoking topic with specific title',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'categories': ['Smoking'],
+      'tags': ['Patient']
+    });
+
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    await tester.ensureVisible(find.text('Smoking'));
+    await tester.pumpAndSettle();
+    await tester.tap(find.text('Smoking'));
+    await tester.pumpAndSettle();
+
+    //expecting both topics to show
+    Finder cardFinder = find.byType(Card);
+    expect(cardFinder, findsNWidgets(2));
+
+    final searchTextField = find.byType(TextField);
+
+    await tester.enterText(searchTextField, 'specific');
+    await tester.pump();
+
+    //smoking topic should be gone
+    expect(cardFinder, findsNWidgets(1));
+
+    expect(find.text('Smoking topic'), findsNothing);
+    expect(find.text('Smoking topic with specific title'), findsOneWidget);
+  });
+
+  testWidgets('Show Post Dialog Test', (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
     // Trigger the _showPostDialog method
     await tester.tap(find.text('Ask a question!'));
     await tester.pumpAndSettle();
-    // Verify that the AlertDialog is displayed
-    //expect(find.byType(AlertDialog), findsOneWidget);
+
+    // Verify that the first AlertDialog is displayed
+    expect(find.byType(AlertDialog), findsOneWidget);
+
     // Enter text into the TextField
     await tester.enterText(find.byType(TextField).last, 'Test question');
+
     // Tap the Submit button
     await tester.tap(find.text('Submit'));
     await tester.pumpAndSettle();
 
+    // Verify that the second AlertDialog is displayed
+    expect(find.text('Message'), findsOneWidget);
+
+    // Tap the OK button to close the second AlertDialog
+    await tester.tap(find.text('OK'));
+    await tester.pumpAndSettle();
+
+    // Verify that both AlertDialogs are closed
+    expect(find.byType(AlertDialog), findsNothing);
+
+    // Verify that the question is added to the Firestore collection
     final QuerySnapshot<Map<String, dynamic>> querySnapshot =
         await firestore.collection("questions").get();
     final List<DocumentSnapshot<Map<String, dynamic>>> documents =
         querySnapshot.docs;
+
     // Check if the collection contains a document with the expected question
     expect(
       documents.any(
@@ -191,7 +431,94 @@ void main() {
       ),
       isTrue,
     );
-    // Verify that the dialog is closed
-    expect(find.byType(AlertDialog), findsNothing);
+  });
+
+  testWidgets('test that shown topics are only of the same role as user',
+      (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+
+    CollectionReference topicCollectionRef = firestore.collection('topics');
+    topicCollectionRef.add({
+      'title': 'test 1',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
+    });
+    topicCollectionRef.add({
+      'title': 'test 2',
+      'description': 'this is a test again',
+      'articleLink': '',
+      'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Parent']
+    });
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    expect(find.text('test 1'), findsOneWidget);
+    expect(find.text('test 2'), findsNothing);
+  });
+
+  testWidgets('test that admin can see all topics',
+      (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+    await auth.createUserWithEmailAndPassword(
+        email: 'admin@tested.org', password: 'Password123!');
+    String uid = auth.currentUser!.uid;
+    await firestore.collection('Users').doc(uid).set({
+      'email': 'admin@tested.org',
+      'firstName': 'James',
+      'lastName': 'Doe',
+      'roleType': 'admin'
+    });
+
+    CollectionReference topicCollectionRef = firestore.collection('topics');
+    topicCollectionRef.add({
+      'title': 'test 1',
+      'description': 'this is a test',
+      'articleLink': '',
+      'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Patient']
+    });
+    topicCollectionRef.add({
+      'title': 'test 2',
+      'description': 'this is a test again',
+      'articleLink': '',
+      'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Parent']
+    });
+    topicCollectionRef.add({
+      'title': 'test 3',
+      'description': 'this is a test again',
+      'articleLink': '',
+      'videoUrl': '',
+      'views': 1,
+      'date': DateTime.now(),
+      'likes': 0,
+      'dislikes': 0,
+      'tags': ['Healthcare Professional']
+    });
+    await tester.pumpWidget(discoveryViewWidget);
+    await tester.pumpAndSettle();
+
+    expect(find.text('test 1'), findsOneWidget);
+    expect(find.text('test 2'), findsOneWidget);
+    expect(find.text('test 3'), findsOneWidget);
   });
 }
