@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth_mocks/firebase_auth_mocks.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
@@ -7,26 +8,30 @@ import 'package:info_hub_app/patient_experience/experience_model.dart';
 
 void main() {
   late FakeFirebaseFirestore firestore;
+  late MockFirebaseAuth auth;
   late CollectionReference topicsCollectionRef;
   late Widget experienceViewWidget;
 
   setUp(() {
     firestore = FakeFirebaseFirestore();
+    auth = MockFirebaseAuth();
     topicsCollectionRef = firestore.collection('experiences');
 
     topicsCollectionRef.add({
       'title': 'Example 1',
       'description': 'Example experience',
+      'userEmail' : 'test@example.org',
       'verified': true
     });
     topicsCollectionRef.add({
       'title': 'Example 2',
       'description': 'Example experience',
+      'userEmail' : 'test2@example.org',
       'verified': false
     });
 
     experienceViewWidget = MaterialApp(
-      home: AdminExperienceView(firestore: firestore),
+      home: AdminExperienceView(firestore: firestore, auth: auth,),
     );
   });
 
@@ -39,17 +44,24 @@ void main() {
     expect(listViewFinder, findsNWidgets(2));
   });
 
-  testWidgets('Displays both verified and unverified experiences',
+  testWidgets('Displays verified experiences',
       (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(experienceViewWidget);
     await tester.pumpAndSettle();
 
-    Finder cardFinder = find.byType(Card);
-    expect(cardFinder, findsNWidgets(2));
-
     expect(find.text('Example 1'), findsOneWidget);
+
+  });
+
+  testWidgets('Displays unverified experiences',
+      (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+    await tester.pumpWidget(experienceViewWidget);
+    await tester.pumpAndSettle();
+
     expect(find.text('Example 2'), findsOneWidget);
+
   });
 
   testWidgets('Button can verify experience correctly',
@@ -99,5 +111,44 @@ void main() {
     expect(experienceList[0].verified, isFalse);
   });
 
-  testWidgets('Share experience works', (WidgetTester tester) async {});
+  testWidgets('Delete button removes an experience from verified list',
+      (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+    await tester.pumpWidget(experienceViewWidget);
+    await tester.pumpAndSettle();
+
+    Finder deleteButton = find.byIcon(Icons.delete).first;
+
+    await tester.ensureVisible(deleteButton);
+    await tester.tap(deleteButton);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(AlertDialog), findsOneWidget);
+
+    await tester.tap(find.text('OK'));
+    await tester.pumpAndSettle();
+
+    expect(find.text('Example 1'), findsNothing);
+  });
+
+  testWidgets('Delete button removes an experience from unverified list',
+      (WidgetTester tester) async {
+    // Build our app and trigger a frame.
+    await tester.pumpWidget(experienceViewWidget);
+    await tester.pumpAndSettle();
+
+    Finder deleteButton = find.byIcon(Icons.delete).last;
+
+    await tester.ensureVisible(deleteButton);
+    await tester.tap(deleteButton);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(AlertDialog), findsOneWidget);
+
+    await tester.tap(find.text('OK'));
+    await tester.pumpAndSettle();
+
+    expect(find.text('Example 2'), findsNothing);
+  });
+
 }
